+++ src/vekna/specs/__init__.py
@@ -1,7 +1,7 @@
 from vekna.specs.constants import (
     ATTENTION_POLL_INTERVAL_SECONDS,
     ATTENTION_WINDOW_STATUS_STYLE,
-    IDLE_TYPING_THRESHOLD_SECONDS,
+    IDLE_THRESHOLD_SECONDS,
     paths_for,
     stem_for_cwd,
     stem_from_tmux_env,
@@ -10,7 +10,7 @@ from vekna.specs.constants import (
 __all__ = [
     "ATTENTION_POLL_INTERVAL_SECONDS",
     "ATTENTION_WINDOW_STATUS_STYLE",
-    "IDLE_TYPING_THRESHOLD_SECONDS",
+    "IDLE_THRESHOLD_SECONDS",
     "paths_for",
     "stem_for_cwd",
     "stem_from_tmux_env",